refactor(world_entry): split cell_dispatch.rs by message family (#244)
cell_dispatch.rs was 865 lines (over the 700-line cap), driven mostly
by the giant match in `handle_cell_message` plus several arms with
inlined wire-emit / DB-touching bodies. Split the inline-heavy arms
into per-family modules so the match itself shrinks back to a thin
"which family handles this variant" dispatch:

  cell_dispatch/mod.rs    — handle_cell_message match (orchestration)
  cell_dispatch/aoi.rs    — entered_aoi, left_aoi, entity_moved,
                            entity_method_call, witness_entity_method,
                            entity_invisible (six AoI packet emitters)
  cell_dispatch/minigame.rs — start_minigame (URL build + onStartMinigame
                              push), minigame_result (onEndMinigame +
                              cell_tx forward)
  cell_dispatch/bandolier.rs — active_slot_update (DB UPDATE + appearance
                                refresh), bandolier_ammo_update (bounds
                                validation + DB writeback)
  cell_dispatch/tests.rs  — existing test suite, unchanged behavior

Variants whose arms were already 1-line wrappers around `handle_*`
helpers in `super::methods::*` stay inline in the match — extracting
them would just add a per-arm hop without changing readability.

Public surface unchanged: `handle_cell_message` is still
`pub(crate)` with the same signature, called from the world-entry
service loop.

All 2 tests pass; clippy clean; cargo fmt applied.
